% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_case_data.R
\name{get_case_data}
\alias{get_case_data}
\title{Create a dataframe of case data, according to the specifications you set.}
\usage{
get_case_data(area_type, area_name, start_date, end_date)
}
\arguments{
\item{area_type}{Specify the name of the area for which you want to get the data.}

\item{area_name}{Specify the area (by name, not code) for which you would like to pull the data.}

\item{start_date}{Specify the first date for which you wish to collect the data. ymd format.}

\item{end_date}{Specify the last date for which you wish to collect the data. ymd format.}
}
\value{
A dataframe containing case data by the specified area type, by date reported and of specimen. Note not all variables are available for all area types.
}
\description{
Makes API call to UK Government coronavirus dashboard and collects data on cases according to the parameters you set. Data will be by nation, region or local authority (upper or lower tier). Note only 1,000 results max returned and that throttling can be implemented if you make too many requests.
}
