% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_hospital_admissions.R
\name{get_hospital_admissions}
\alias{get_hospital_admissions}
\title{Create a dataframe of hospital admissions data, according to the specifications you set.}
\usage{
get_hospital_admissions(region, start_date, end_date)
}
\arguments{
\item{region}{Specify the region for which you want to get the data.}

\item{start_date}{Specify the first date for which you wish to collect the data. ymd format.}

\item{end_date}{Specify the last date for which you wish to collect the data. ymd format.}
}
\value{
A dataframe containing hospital admissions data by NHS region: new admissions and cumulative admissions. Only a maximum of 1,000 results will be returned.
}
\description{
Makes API call to UK Government coronavirus dashboard and collects data on admissions according to the parameters you set. Data will be by NHS region and only for English regions. Note only 1,000 results max returned.
}
